+++ generator_test.go
@@ -1,11 +1,19 @@
 package gemini
 
 import (
+	"reflect"
+	"sync/atomic"
 	"testing"
-	"time"
+
+	"go.uber.org/zap"
 )
 
 func TestGenerator(t *testing.T) {
+	table := &Table{
+		Name:          "tbl",
+		PartitionKeys: createPkColumns(1, "pk"),
+	}
+	var current uint64
 	cfg := &GeneratorsConfig{
 		Partitions: PartitionRangeConfig{
 			MaxStringLength: 10,
@@ -14,31 +22,21 @@ func TestGenerator(t *testing.T) {
 			MinBlobLength:   0,
 		},
 		Size:             1,
-		PkBufferSize:     10000,
 		PkUsedBufferSize: 10000,
-		Table: &Table{
-			Name:          "tbl",
-			PartitionKeys: createPkColumns(1, "pk"),
+		DistributionSize: 1000,
+		DistributionFunc: func() uint64 {
+			return atomic.LoadUint64(&current)
 		},
 	}
-	generators := NewGenerator(cfg)
+	logger, _ := zap.NewDevelopment()
+	generators := NewGenerator(table, cfg, logger)
 	source := generators.Get(0)
-
-	time.Sleep(time.Second)
-
-	if size := len(source.newValues); size != 10000 {
-		t.Errorf("expected %d pks got %d", 10000, size)
-	}
-	if size := len(source.oldValues); size != 0 {
-		t.Errorf("expected %d spent pks got %d", 0, size)
-	}
-
-	generators.Stop()
-	cnt := 0
-	for range source.newValues {
-		cnt++
-	}
-	if cnt < 10000 {
-		t.Errorf("expected at least %d pks after stop got %d", 10000, cnt)
+	for i := uint64(0); i < cfg.DistributionSize; i++ {
+		atomic.StoreUint64(&current, i)
+		v, _ := source.Get()
+		n, _ := source.Get()
+		if !reflect.DeepEqual(v, n) {
+			t.Errorf("expected %v, got %v", v, n)
+		}
 	}
 }